Added admin order approval and sending pdf order detail to email

+++ order/models.py
@@ -25,7 +25,7 @@ class Order(models.Model):
     end_date = models.CharField(max_length=20)
     approval = models.BooleanField(default=False)
     finished = models.BooleanField(default=False)
-    total_price = models.IntegerField(null=True, blank=True)
+    total_price = models.FloatField(null=True, blank=True)
     comment = models.CharField(max_length=100, null=True, blank=True)
     user = models.ForeignKey(settings.AUTH_USER_MODEL, on_delete=models.CASCADE,
                              related_name='orders')
